usr/requirements.txt
speedtest-cli
rarfile
pychromecast
zeroconf
PyQt5
os
sys
subprocess
asyncio
pyatv
pip install py7zr

sudo apt-get install python3-PyQt5.QtMultimedi // mierda no sirve
python3-vlc// sirve pero crashea con PyQt5 pero da bien con otras GUI

pip install python-mpv // mierda  mas bugeada
PyOpenGL PyOpenGL_accelerate
pip install pyatv --break-system-package    pip install pyatv==0.6.0
pip install PyQt5 PyQtWebEngine --break-system-packages sudo apt install python3-pyqt5.qtwebengine
